remove dynamic logging
